Several fixes
Featured photo loads once only. Does not update with scrolling.

Updated loaders with new sliver framework

Changed name of dartfiles to follow a standard

deleted chewy images from assets/images

broke out loaders to own file
